+++ src/main/java/org/openurp/edu/base/model/Squad.java
@@ -44,7 +44,7 @@ import org.openurp.edu.base.code.model.StdType;
 @Entity(name = "org.openurp.edu.base.model.Squad")
 @Cacheable
 @Cache(region = "openurp.base", usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
-public class Squad extends EduLevelBasedObject<Integer> {
+public class Squad extends EduLevelBasedObject<Long> {
 
   private static final long serialVersionUID = 6467000522483394459L;
 
@@ -120,11 +120,11 @@ public class Squad extends EduLevelBasedObject<Integer> {
   public Squad() {
   }
 
-  public Squad(Integer id) {
+  public Squad(Long id) {
     super(id);
   }
 
-  public Squad(Integer id, String code, String name) {
+  public Squad(Long id, String code, String name) {
     super(id);
     this.code = code;
     this.name = name;
